backend/requirements.txt
pymongo
flask
flask-cors
python-dotenv
google-genai
langgraph